requirements.txt
numpy==2.1.1
pygame==2.6.0
pygame-widgets==1.1.5
PyYAML==6.0.2
scipy==1.14.1
screeninfo==0.8.1
